% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat2}
\alias{dat2}
\title{Simulated data from the longitudinal study}
\format{
\subsection{\code{dat2}}{

A data frame with 500 rows and 15 columns:
\describe{
\item{x1, x2, x3, x4}{The original continuous baseline covariates that
are used to generate the non-linear transformed covariates}
\item{x5}{The original binary baseline covariate}
\item{z1, z2, z3, z4}{The transformed continuous baseline covariates that
are used to generate the data}
\item{z5}{The transformed binary baseline covariate}
\item{a}{The treatment assignment.}
\item{r1}{The response status at the first visit, where \code{r1 = 1} indicates
the subject is observed at the first visit, and \verb{r1 = 0`` implies the subject is missing at the first visit.\} \\item\{r2\}\{The response status at the second visit, where }r2 = 1\verb{indicates the subject is observed at the second visit, and}r2 = 0` implies the
subject is missing at the second visit.}
\item{y1}{The outcome at the first visit. Note that "NA" indicates that
the subject is missing at the first visit, thus has no outcomes.}
\item{y2}{The outcome at the second visit. Note that "NA" indicates that
the subject is missing at the second visit, thus has no outcomes.}
}
}
}
\usage{
dat2
}
\description{
The simulated dataset for the longitudinal study that includes two follow-up
visits.
}
\keyword{datasets}
