% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datHAMD}
\alias{datHAMD}
\title{An antidepressant trial data}
\format{
\subsection{\code{datHAMD}}{

A data frame with 196 rows and 14 columns:
\describe{
\item{id}{The original continuous baseline covariates that
are used to generate the non-linear transformed covariates.}
\item{trt}{The treatment assignment.}
\item{poolinv}{The categorical variable indicating the investigation sites.}
\item{base}{The baseline variable.}
\item{y1,y2,y3,y4,y5}{The relative change from baseline at the k-th
follow-up visit, with \code{NA} indicates the missing outcome and k = 1,...,5.}
\item{r1,r2,r3,r4,r5}{The response status at the k-th follow-up visit,
where \code{rk = 1} indicates the subject is observed at the k-th visit, and
\code{rk = 0} implies the subject is missing at the k-th visit, for k = 1,...,5.}
}
}
}
\source{
\url{https://www.lshtm.ac.uk/research/centres-projects-groups/missing-data#dia-missing-data}
}
\usage{
datHAMD
}
\description{
The data targets an antidepressant trial conducted under the Auspices of the
Drug Information Association. It was prepared by Mallinckrodt et al.(2014).
To illustrate the usage of our proposed estimator, we delete the individual
with intermittent missingness and only focus on the individuals with a
monotone missingness pattern, and we also delete three individuals with the
unobserved investigation site numbers. The reorganized dataset contain 99
subjects in the control group and 97 subjects in the treatment group. After
data preprocessing, 39 participants in the control group and 30 participants
in the treatment group suffered from monotone missingness.
}
\keyword{datasets}
