% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{jtr.1time}
\alias{jtr.1time}
\title{Obtain the proposed estimator in the cross-sectional study}
\usage{
jtr.1time(formula.ps, formula.rp, formula.om, data, mat.cal = NULL, type)
}
\arguments{
\item{formula.ps}{a list of formulas for fitting the propensity score model}

\item{formula.rp}{a list of formulas for fitting the response probability
model}

\item{formula.om}{a list of formulas for fitting the outcome mean model}

\item{data}{the data that include the binary treatment assignment, where the
active treatment group should be encoded as \code{1}, and the control group should be
encoded as \code{0}, the binary response status, where \code{1} indicates the subject
is observed and \code{0} indicates the subject is missing, the outcome, and the
fully observed baseline covariates.}

\item{mat.cal}{the matrix for calibration, should have the same subject order
as the original data.}

\item{type}{the type of estimate the user want to obtain. Available types
include:
\itemize{
\item \code{tr}: the triply robust estimator
\item \code{tr.norm}: the triply robust estimator with normalization
\item \code{tr.cal}: the triply robust estimator with calibration
\item \code{psom}: the ps-om estimator
\item \code{psom.norm}: the ps-om estimator with normalization
\item \code{psrp}: the ps-rp estimator
\item \code{psrp.norm}: the ps-rp estimator with normalization
\item \code{rpom}: the rp-om estimator
}}
}
\value{
The point estimate
}
\description{
This function derives the estimators for the average treatment effect (ATE)
under jump-to-reference (J2R) in the cross-sectional study. The resulting ATE
estimators consists of five simple estimators (rp-om, ps-om, ps-om-N, ps-rp,
and ps-rp-N) and three triply robust estimators (tr, tr-N, and tr-C)
motivated by the efficient influence function (EIF). Also, EIF-based
variance estimators are provided for the three EIF-based estimators. The
nuisance functions are estimated by parametric models.
}
\examples{
data <- dat1
formula.om <- y ~ z1 + z2 + z3 + z4 + z5 + a +
  a:(z1 + z2 + z3 + z4 + z5)
formula.ps <- a ~ z1 + z2 + z3 + z4 + z5
formula.rp <- r ~ z1 + z2 + z3 + z4 + z5 + a +
  a:(z1 + z2 + z3 + z4 + z5)
mat.cal <- data.matrix(data[,paste0("z", 1:5)]) # calibration matrix
res.1time <- jtr.1time(formula.ps = formula.ps,
                       formula.rp = formula.rp,
                       formula.om = formula.om,
                       data = data, mat.cal = mat.cal,
                       type = c("tr", "tr.cal", "rpom"))
res.1time
}
