% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{jtr.longi}
\alias{jtr.longi}
\title{Obtain the proposed estimator in the longitudinal study}
\usage{
jtr.longi(
  formula.ps,
  formula.rp,
  formula.om,
  formula.pm,
  data,
  mat.cal = NULL,
  type
)
}
\arguments{
\item{formula.ps}{a list of formulas for fitting the propensity score model.}

\item{formula.rp}{a list of formulas for fitting the response probability
model.}

\item{formula.om}{a list of formulas for fitting the outcome mean model.}

\item{formula.pm}{a list of formulas for fitting the pattern mean model.
Users can input any variables in the data as the response variables, since
the response variables in the pattern mean formulas are recalculated by the
estimated outcome mean and response probability.}

\item{data}{the data that include the binary treatment assignment, where the
active treatment group should be encoded as \code{1}, and the control group should be
encoded as \code{0}, the binary response status, where \code{1} indicates the subject
is observed and \code{0} indicates the subject is missing, the outcome, and the
fully observed baseline covariates.}

\item{mat.cal}{the matrix for calibration, should have the same subject order
as the original data.}

\item{type}{the type of estimate the user want to obtain. Available types
include:
\itemize{
\item \code{mr}: the multiply robust estimator
\item \code{mr.norm}: the triply robust estimator with normalization
\item \code{psom}: the ps-om estimator
\item \code{psom.norm}: the ps-om estimator with normalization
\item \code{psrp}: the ps-rp estimator
\item \code{psrp.norm}: the ps-rp estimator with normalization
\item \code{rppm}: the rp-pm estimator
}}
}
\value{
The point estimate
}
\description{
This function derives the estimators for the average treatment effect (ATE)
under jump-to-reference (J2R) in the longitudinal study. The resulting ATE
estimators consists of five simple estimators (rp-pm, ps-om, ps-om-N, ps-rp,
and ps-rp-N) and two multiply robust estimators (mr and mr-N) motivated
by the efficient influence function (EIF). Also, EIF-based variance
estimators are provided for the two EIF-based estimators. The
nuisance functions are estimated by generalized additive models (GAMs).
Users can select the type of smoothness used in GAMs.
}
\examples{
data <- dat2
formula.om <- list()
formula.om[[2]] <- y2 ~ gam::s(x1) + gam::s(x2) + gam::s(x3) + gam::s(x4) +
x5 + gam::s(y1)
formula.om[[1]] <- y1 ~ gam::s(x1) + gam::s(x2) + gam::s(x3) + gam::s(x4) + x5
formula.ps <- list()
formula.ps[[2]] <- a ~ gam::s(x1) + gam::s(x2) + gam::s(x3) + gam::s(x4) +
x5  + gam::s(y1)
formula.ps[[1]] <- a ~ gam::s(x1) + gam::s(x2) + gam::s(x3) + gam::s(x4) + x5
formula.rp <- list()
formula.rp[[2]] <- r2 ~ gam::s(x1) + gam::s(x2) + gam::s(x3) + gam::s(x4) +
x5 + gam::s(y1)
formula.rp[[1]] <- r1 ~ gam::s(x1) + gam::s(x2) + gam::s(x3) + gam::s(x4) + x5
formula.pm <- list()
formula.pm[[2]] <- y2 ~ gam::s(x1) + gam::s(x2) + gam::s(x3) + gam::s(x4) +
x5 + gam::s(y1)
formula.pm[[1]] <- y1 ~ gam::s(x1) + gam::s(x2) + gam::s(x3) + gam::s(x4) + x5
mat.cal <- data.matrix(data[,paste0("z", 1:5)]) # calibration matrix
res.longi <- jtr.longi(formula.ps = formula.ps,
                       formula.rp = formula.rp,
                       formula.om = formula.om,
                       formula.pm = formula.om,
                       data = data,
                       type = c("mr", "mr.norm", "rppm"))
res.longi
}
