% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_1time.R, R/sim_longi.R
\name{main}
\alias{main}
\title{Main function to get both point and variance estimates}
\usage{
main(seed, n, k, alpha, gamma1, gamma2, B)

main(seed, n, k, alpha, gamma1, gamma2, B)
}
\arguments{
\item{seed}{a specific seed number}

\item{n}{the sample size}

\item{k}{the dimension of baseline covariates}

\item{alpha}{coefficients of the propensity score}

\item{gamma1}{coefficients of the response probability for R1}

\item{gamma2}{coefficients of the response probability for R2}

\item{B}{the number of bootstrap replicates}

\item{gamma}{coefficients of the response probability}
}
\value{
the estimated point and variance estimates

the estimated point and variance estimates
}
\description{
Main function to get both point and variance estimates

Main function to get both point and variance estimates
}
