% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_1time.R, R/sim_longi.R
\name{model_est}
\alias{model_est}
\title{Obtain the proposed estimator in the cross-sectional study}
\usage{
model_est(dat, k)

model_est(dat, k)
}
\arguments{
\item{dat}{the simulated dataset}

\item{k}{the dimension of baseline covariates}
}
\value{
The point estimates of the eight proposed estimators

The point estimates of the eight proposed estimators
}
\description{
Obtain the proposed estimator in the cross-sectional study

Obtain the proposed estimator in the longitudinal study
}
