% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_1time.R, R/sim_longi.R
\name{nonpara_fn}
\alias{nonpara_fn}
\title{Nonparametric bootstrap to obtain the variance estimates}
\usage{
nonpara_fn(dat, B, point_est)

nonpara_fn(dat, B, point_est)
}
\arguments{
\item{dat}{the simulated dataset}

\item{B}{the number of bootstrap replicates}

\item{point_est}{a vector of the point estimates of the eight proposed
estimators}

\item{psrpom_est}{a vector of the point estimates of the eight proposed
estimators when three models are correctly specified}

\item{psom_est}{a vector of the point estimates of the eight proposed
estimators when (ps, om) are correctly specified}

\item{ps_est}{a vector of the point estimates of the eight proposed
estimators when (ps) is correctly specified}

\item{rpom_est}{a vector of the point estimates of the eight proposed
estimators when (rp, om) are correctly specified}

\item{rp_est}{a vector of the point estimates of the eight proposed
estimators when (rp) is correctly specified}

\item{psrp_est}{a vector of the point estimates of the eight proposed
estimators when (ps, rp) are correctly specified}

\item{om_est}{a vector of the point estimates of the eight proposed
estimators when (om) is correctly specified}

\item{none_est}{a vector of the point estimates of the eight proposed
estimators when none of the modesl are correctly specified}
}
\value{
the estimated variation of the estimators using nonparmetric
bootstrap, symmetric-t bootstrap, and bootstrap percentile

the estimated variation of the estimators using nonparmetric
bootstrap, symmetric-t bootstrap, and bootstrap percentile
}
\description{
Nonparametric bootstrap to obtain the variance estimates

Nonparametric bootstrap to obtain the variance estimates
}
