% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_1time.R, R/sim_longi.R
\name{sim_dat}
\alias{sim_dat}
\title{Generate the data from a simulated longitudinal study}
\usage{
sim_dat(n, k, alpha, gamma1, gamma2)

sim_dat(n, k, alpha, gamma1, gamma2)
}
\arguments{
\item{n}{the sample size}

\item{k}{the dimension of baseline covariates}

\item{alpha}{coefficients of the propensity score}

\item{gamma1}{coefficients of the response probability for R1}

\item{gamma2}{coefficients of the response probability for R2}

\item{gamma}{coefficients of the response probability}
}
\value{
A data frame

A data frame
}
\description{
Generate the data from a simulated longitudinal study

Generate the data from a simulated longitudinal study
}
