% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{jtr.1time}
\alias{jtr.1time}
\title{Obtain the proposed estimator in the cross-sectional study}
\usage{
jtr.1time(formula.ps, formula.rp, formula.om, data, mat.cal = NULL, type)
}
\arguments{
\item{formula.ps}{the matrix of all covariates included in the propensity
score model, with the dimension: n*p1, n: the sample size, p1: dimension of the
covariates.}

\item{formula.rp}{the matrix of all covariates included in the response status
model, with the dimension: n*p2, n: the sample size, p2: dimension of the
covariates.}

\item{formula.om}{the matrix of all covariates included in the outcome mean
model, with the dimension: n*p3, n: the sample size, p3: dimension of the
covariates.}

\item{data}{the data that include the binary treatment assignment, where the
active treatment group should be encoded as \code{1}, and the control group should be
encoded as \code{0}, the binary response status, where \code{1} indicates the subject
is observed and \code{0} indicates the subject is missing, the outcome, and the
fully observed baseline covariates.}

\item{mat.cal}{the matrix for calibration, should have the same subject order
as the original data.}

\item{type}{the type of estimate the user want to obtain. Available types
include:
\itemize{
\item \code{tr}: the triply robust estimator with calibration
\item \code{tr.norm}: the triply robust estimator with normalization
\item \code{tr.cal}: the triply robust estimator with calibration
\item \code{psom}: the ps-om estimator
\item \code{psom.norm}: the ps-om estimator with normalization
\item \code{psrp}: the ps-rp estimator
\item \code{psrp.norm}: the ps-rp estimator with normalization
\item \code{rpom}: the rp-om estimator
}}
}
\value{
The point estimate
}
\description{
This function implements the proposed RPSFT + Cox method and returns an
estimate of the treatment effect on the overall survival. It can be used if
investigators have sufficient knowledge of the PDT effects and the survival
time during the PDT period.
}
\examples{
data <- dat1
formula.om <- y ~ z1 + z2 + z3 + z4 + z5 + a +
  a:(z1 + z2 + z3 + z4 + z5)
formula.ps <- a ~ z1 + z2 + z3 + z4 + z5
formula.rp <- r ~ z1 + z2 + z3 + z4 + z5 + a +
  a:(z1 + z2 + z3 + z4 + z5)
mat.cal <- data.matrix(data[,paste0("z", 1:5)]) # calibration matrix
res.1time <- jtr.1time(formula.ps = formula.ps,
                       formula.rp = formula.rp,
                       formula.om = formula.om,
                       data = data, mat.cal = mat.cal,
                       type = c("tr", "tr.cal", "rpom"))
res.1time
}
